requirements.txt

fastapi
uvicorn[standard]
google-cloud-aiplatform
python-multipart
pdfplumber
python-docx
streamlit==1.35.0
google-cloud-storage
pandas==2.2.2
db-dtypes==1.2.0
langchain==0.2.1
langchain-groq==0.1.4
langchain-core==0.2.3
python-dotenv==1.0.1
joblib==1.4.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
unidecode==1.3.8
pandas-gbq==0.23.2
pyarrow==16.1.0
unidecode==1.3.8
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage
nltk==3.9.1
pytest==8.3.2
pytest-cov==5.0.0
google-genai==1.38.0
google-generativeai==0.8.5
httpx==0.28.1

